fix(objects/revit): Merged brep changes into shared project
